Fix Node.js modules usage in client-side code
