+++ next.config.js
@@ -1,0 +1,21 @@
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  reactStrictMode: true,
+  webpack: (config, { isServer }) => {
+    if (!isServer) {
+      // Don't resolve 'fs', 'path', etc. on the client
+      config.resolve.fallback = {
+        fs: false,
+        path: false,
+        crypto: false,
+        stream: false,
+        util: false,
+        os: false,
+        child_process: false,
+      };
+    }
+    return config;
+  },
+}
+
+export default nextConfig;
